+++ backend/src/shot.go
@@ -7,12 +7,12 @@ import (
 type Shot struct {
 	id    uint64
 	owner *Player
-	xPos  int
-	yPos  int
+	xPos  float64
+	yPos  float64
 	angle int
 }
 
 func (s *Shot) move() {
-	s.xPos = int(globalShotSpeed*math.Cos(float64(s.angle)*math.Pi/180.0)) + s.xPos
-	s.yPos = int(globalShotSpeed*math.Sin(float64(s.angle)*math.Pi/180.0)) + s.yPos
+	s.xPos = globalShotSpeed*math.Cos(float64(s.angle)*math.Pi/180.0) + s.xPos
+	s.yPos = globalShotSpeed*math.Sin(float64(s.angle)*math.Pi/180.0) + s.yPos
 }
